Use Doctrine for DB access in web app.
