--- src/App/src/Db/Table/Message.php
@@ -1,52 +1,0 @@
-<?php
-/**
- * Message table gateway
- *
- * PHP version 7
- *
- * Copyright (C) Villanova University 2020.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- *
- * @category VuOwma
- * @package  Database
- * @author   Demian Katz <demian.katz@villanova.edu>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://github.com/FalveyLibraryTechnology/VuOwma/
- */
-namespace App\Db\Table;
-
-use Laminas\Db\Adapter\Adapter;
-
-/**
- * Message table gateway
- *
- * @category VuOwma
- * @package  Database
- * @author   Demian Katz <demian.katz@villanova.edu>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://github.com/FalveyLibraryTechnology/VuOwma/
- */
-class Message extends \Laminas\Db\TableGateway\TableGateway
-{
-    /**
-     * Constructor
-     *
-     * @param Adapter $adapter Database adapter
-     */
-    public function __construct(Adapter $adapter)
-    {
-        parent::__construct('messages', $adapter);
-    }
-}
