+++ src/App/src/Handler/DoctrineAwareFactory.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Batch table gateway
+ * Factory for handlers depending on Doctrine.
  *
  * PHP version 7
  *
- * Copyright (C) Villanova University 2020.
+ * Copyright (C) Villanova University 2021.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -20,33 +20,50 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  *
  * @category VuOwma
- * @package  Database
+ * @package  Handlers
  * @author   Demian Katz <demian.katz@villanova.edu>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://github.com/FalveyLibraryTechnology/VuOwma/
  */
-namespace App\Db\Table;
+namespace App\Handler;
 
-use Laminas\Db\Adapter\Adapter;
+use Interop\Container\ContainerInterface;
 
 /**
- * Batch table gateway
+ * Factory for handlers depending on Doctrine.
  *
  * @category VuOwma
- * @package  Database
+ * @package  Handlers
  * @author   Demian Katz <demian.katz@villanova.edu>
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://github.com/FalveyLibraryTechnology/VuOwma/
  */
-class Batch extends \Laminas\Db\TableGateway\TableGateway
+class DoctrineAwareFactory
 {
     /**
-     * Constructor
+     * Create an object
      *
-     * @param Adapter $adapter Database adapter
+     * @param ContainerInterface $container     Service manager
+     * @param string             $requestedName Service being created
+     * @param null|array         $options       Extra options (optional)
+     *
+     * @return object
+     *
+     * @throws ServiceNotFoundException if unable to resolve the service.
+     * @throws ServiceNotCreatedException if an exception is raised when
+     * creating a service.
+     * @throws ContainerException if any other error occurs
      */
-    public function __construct(Adapter $adapter)
-    {
-        parent::__construct('batches', $adapter);
+    public function __invoke(
+        ContainerInterface $container,
+        $requestedName,
+        array $options = null
+    ) {
+        if (!empty($options)) {
+            throw new \Exception('Unexpected options passed to factory.');
+        }
+        return new $requestedName(
+            $container->get(\Doctrine\ORM\EntityManager::class)
+        );
     }
 }
